moving lock to another file, cleanup
