+++ lib/features/interactive.js
@@ -7,48 +7,12 @@ import https from 'https';
 import path from 'path';
 import Promise from 'promise';
 import slugid from 'slugid';
+import SharedFileLock from '../shared_file_lock';
 import url from 'url';
 let debug = Debug('docker-worker:features:interactive');
 
 //Number of attempts to find a port before giving up
 const MAX_RANDOM_PORT_ATTEMPTS = 20;
-
-class Semaphore {
-  constructor(lockFile) {
-    this.count = 0;
-    this.lockFile = lockFile;
-  }
-
-  //acquires a lock, at >=1 locks it will flock the lockfile
-  async acquire() {
-    if(this.count === 0) {
-      let err = await Promise.denodeify(fs.flock)(this.lockFile, 'shnb');
-      if(err) {
-        debug('couldn\'t acquire lock, this is probably bad');
-        debug(err);
-      }
-    }
-    this.count += 1;
-    debug('acquire; count is %s', this.count);
-  }
-
-  //releases a lock after some delay, at 0 locks it will unlock the lockfile
-  async release(delay = 0) {
-    if(delay > 0) {
-      return(setTimeout(() => {this.release()}, delay));
-    }
-    assert(this.count > 0, "Internal error");
-    this.count -= 1;
-    if(this.count === 0) {
-      let err = await Promise.denodeify(fs.flock)(this.lockFile, 'un');
-      if(err) {
-        debug('couldn\'t acquire lock, this is probably bad');
-        debug(err);
-      }
-    }
-    debug('released; count is %s', this.count);
-  }
-}
 
 export default class WebsocketServer {
   constructor () {
@@ -130,7 +94,7 @@ export default class WebsocketServer {
     });
 
     //set expiration stuff
-    this.semaphore = new Semaphore(this.lockFile);
+    this.semaphore = new SharedFileLock(this.lockFile);
     this.semaphore.acquire();
     this.semaphore.release(task.runtime.interactive.maxTime * 1000);
     this.server.on('session added', () => {
