+++ lib/task.js
@@ -267,7 +267,7 @@ export default class Task {
       //TODO: test with things that aren't undefined
       let oldEntrypoint = (await this.runtime.docker.getImage(config.image).inspect()).Entrypoint;
       if(typeof oldEntrypoint === 'string') {
-        oldEntrypoint = [oldEntrypoint];
+        oldEntrypoint = ['/bin/sh', '-c', oldEntrypoint];
       } else if(typeof oldEntrypoint === 'undefined') {
         oldEntrypoint = [];
       }
